requirements.txt
about-time==3.1.1
alive-progress==2.0.0
beautifulsoup4==4.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
click-plugins==1.1.1
colorama==0.4.4
dnspython==2.0.0
ExifRead==2.3.2
future==0.18.2
google==3.0.0
grapheme==0.6.0
idna==3.2
install==1.3.4
ipwhois==1.2.0
lxml==4.6.3
nslookup==1.4.0
progress==1.6
progressbar2==3.53.1
pycrtsh==0.3.5
PyPDF2==1.26.0
python-dateutil==2.8.2
python-utils==2.5.6
python-whois==0.7.3
requests==2.26.0
shodan==1.25.0
six==1.16.0
soupsieve==2.2.1
tqdm==4.62.2
urllib3==1.26.6
XlsxWriter==3.0.1
